Update improve code
